tracker 3378757: pdb support on Windows

+++ clewn/debugger.py
@@ -559,7 +559,8 @@ class Debugger:
         Annotations are not deleted.
 
         """
-        self.__nbsock.remove_all()
+        if self.__nbsock:
+            self.__nbsock.remove_all()
 
     def get_lnum_list(self, pathname):
         """Return a list of line numbers of all enabled breakpoints in a
@@ -721,7 +722,7 @@ class Debugger:
         info('enter netbeans_detach')
         self.started = False
         self.closed = True
-        if self.__nbsock.connected:
+        if self.__nbsock and self.__nbsock.connected:
             self.remove_all()
             msg = 'DETACH'
             info('sending netbeans message \'%s\'', msg)
@@ -893,7 +894,7 @@ class Debugger:
 
         """
         if not self.__nbsock:
-            return
+            return LOOP_TIMEOUT
         timeout = LOOP_TIMEOUT
         if not self._jobs_enabled and self.__nbsock.ready:
             self._jobs_enabled = True
